feat: restructure GitHub commands under `floo apps github` subcommand
Restructure connect/disconnect from flat `floo apps connect` to nested
`floo apps github connect/disconnect/status`. Key changes:

- New GitHubCommands enum with Connect (positional repo), Disconnect, Status
- Move connect/disconnect logic to dedicated commands/github.rs module
- Add `status` subcommand calling GET /v1/apps/{app_id}/github
- Handle GITHUB_REPO_NOT_IN_INSTALLATION error (opens settings page)
- Optional --app flag resolves from floo.app.toml via resolve_app_from_config
- Update references in releases.rs and skills/floo.md
